refactor: convert to PluginSystemController
