+++ packages/plugin-uploader/src/pages/PluginSystemPageInterface.ts
@@ -2,7 +2,7 @@ import type { BoundMessage } from "../messages";
 import type { Page } from "puppeteer";
 import type { Lang } from "../lang";
 
-export interface PluginSystemPageBase {
+export interface PluginSystemPageInterface {
   readyForImportButton(page: Page, m: BoundMessage): Promise<void>;
   upload(page: Page, pluginPath: string, lang: Lang): Promise<void>;
 }
